requirements.txt
tensorflow-cpu
fastapi
uvicorn
gunicorn
python-multipart
pillow
matplotlib
numpy
